NWI 4.0.2

+++ save_group.php
@@ -57,11 +57,7 @@ else
 }
 
 // Update row
-$database->query(sprintf(
-    "UPDATE `%smod_news_img_groups` SET `title` = '$title', `active` = '$active' " .
-    "WHERE `group_id`=%d",
-    TABLE_PREFIX, intval($group_id)
-));
+$database->query("UPDATE `".TABLE_PREFIX."mod_news_img_groups` SET `title` = '$title', `active` = '$active' WHERE `group_id` = '$group_id'");
 
 // Check if the user uploaded an image or wants to delete one
 if(isset($_FILES['image']['tmp_name']) AND $_FILES['image']['tmp_name'] != '')
@@ -73,13 +69,13 @@ if(isset($_FILES['image']['tmp_name']) AND $_FILES['image']['tmp_name'] != '')
 	$file4=substr($filename, -4, 4);
 	if(($file4 != '.jpg')and($file4 != '.JPG')and($file4 != '.png')and($file4 != '.PNG') and ($file4 !='jpeg') and ($file4 != 'JPEG'))
     {
-		$admin->print_error($MESSAGE['GENERIC']['FILE_TYPE'].' JPG (JPEG) or PNG a');
+		$admin->print_error($MESSAGE['GENERIC']['FILE_TYPES'].' JPG, JPEG, PNG', WB_URL.'/modules/news_img/modify_group.php?page_id='.$page_id.'&section_id='.$section_id.'&group_id='.$admin->getIDKEY($group_id));
 	} elseif(
 	(($_FILES['image']['type']) != 'image/jpeg' AND mime_content_type($_FILES['image']['tmp_name']) != 'image/jpg')
 	and
 	(($_FILES['image']['type']) != 'image/png' AND mime_content_type($_FILES['image']['tmp_name']) != 'image/png')
 	){
-		$admin->print_error($MESSAGE['GENERIC']['FILE_TYPE'].' JPG (JPEG) or PNG b');
+		$admin->print_error($MESSAGE['GENERIC']['FILE_TYPES'].' JPG, JPEG, PNG', WB_URL.'/modules/news_img/modify_group.php?page_id='.$page_id.'&section_id='.$section_id.'&group_id='.$admin->getIDKEY($group_id));
 	}
 	// Make sure the target directory exists
 	make_dir(WB_PATH.MEDIA_DIRECTORY.'/.news_img');
@@ -96,12 +92,14 @@ if(isset($_FILES['image']['tmp_name']) AND $_FILES['image']['tmp_name'] != '')
     {
 		// Resize the image
 		$thumb_location = WB_PATH.MEDIA_DIRECTORY.'/.news_img/thumb'.$group_id.'.jpg';
-        if (list($w, $h) = getimagesize($new_filename)) {
+        if (extension_loaded('gd') && list($w, $h) = getimagesize($new_filename)) {
             if ($w>$previewwidth || $h>$previewheight) {
                 mod_nwi_image_resize($new_filename, $thumb_location, $previewwidth, $previewheight, $fetch_settings['crop_preview']);
                 unlink($new_filename);
                 rename($thumb_location,$new_filename);
             }
+        } else {
+            copy($new_filename,$thumb_location);
         }
 	}
 }
